+++ ParabolicFunction/Models/UserData.cs
@@ -11,11 +11,11 @@ namespace ParabolicFunction.Models
     {
         [Key]
         public int UserDataId { get; set; }
-        public int RangeFrom { get; set; }
-        public int RangeTo { get; set; }
-        public int Step { get; set; }
-        public int A { get; set; }
-        public int B { get; set; }
-        public int C { get; set; }
+        public double RangeFrom { get; set; }
+        public double RangeTo { get; set; }
+        public double Step { get; set; }
+        public double A { get; set; }
+        public double B { get; set; }
+        public double C { get; set; }
     }
 }
